% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cmprsk_cut}
\alias{cmprsk_cut}
\title{Цензурирование данных с конкурирующими рисками}
\usage{
cmprsk_cut(
  date_end,
  date_start,
  cmprsk_date,
  status,
  cens_time,
  cutoff_time,
  status_name,
  cmprsk_name,
  data
)
}
\arguments{
\item{date_end}{Дата окончания наблюдения или события (в формате Date).}

\item{date_start}{Дата начала наблюдения (в формате Date).}

\item{cmprsk_date}{Дата наступления конкурирующего риска (в формате Date).}

\item{status}{Статус основного события (0 - цензурировано, 1 - событие произошло).}

\item{cens_time}{Единица времени для цензурирования (число).}

\item{cutoff_time}{Время отсечки для цензурирования (число).}

\item{status_name}{Название основного события (строка, например, "Рецидив").}

\item{cmprsk_name}{Название конкурирующего риска (строка, например, "Смерть").}

\item{data}{Датафрейм, содержащий исходные данные.}
}
\value{
Датафрейм с добавленными колонками: \verb{<cmprsk_name>_status_<cutoff_time>} и \verb{<cmprsk_name>_time_<cutoff_time>}.
}
\description{
Функция выполняет цензурирование данных о конкурирующих рисках на основе заданного времени отсечки.
}
\examples{
# Пример использования функции cmprsk_cut
data <- data.frame(
  date_start = as.Date("2020-01-01"),
  date_end = as.Date("2021-06-01"),
  cmprsk_date = as.Date(c(NA, "2021-01-01", NA)),
  status = c(1, 0, 1)
)
result <- cmprsk_cut(
  date_end = data$date_end,
  date_start = data$date_start,
  cmprsk_date = data$cmprsk_date,
  status = data$status,
  cens_time = 30,         # Время измеряется в месяцах
  cutoff_time = 12,       # 12 месяцев
  status_name = "Рецидив",
  cmprsk_name = "Смерть",
  data = data
)
}
